requirement.txt
torch
numpy
imgui
pyopengl
pygame
numexpr
typeguard
jsonpickle
yaml